# Improve performance of Include and Exclude functions