--- a/tsdb/engine/tsm1/encoding.gen.go.tmpl
+++ b/tsdb/engine/tsm1/encoding.gen.go.tmpl
@@ -85,16 +85,65 @@ func (a {{.Name}}Values) Deduplicate() {{.Name}}Values {
 
 //  Exclude returns the subset of values not in [min, max]
 func (a {{.Name}}Values) Exclude(min, max int64) {{.Name}}Values {
-	var i int
-	for j := 0; j < len(a); j++ {
-		if a[j].UnixNano() >= min && a[j].UnixNano() <= max {
-			continue
-		}
-
-		a[i] = a[j]
-		i++
+	rmin, rmax := a.FindRange(min, max)
+	if rmin == -1 && rmax == -1 {
+		return a
 	}
-	return a[:i]
+
+	// a[rmin].UnixNano() ≥ min
+	// a[rmax].UnixNano() ≥ max
+
+	if rmax < len(a) {
+		if a[rmax].UnixNano() == max {
+			rmax++
+		}
+		rest := len(a)-rmax
+		if rest > 0 {
+			b := a[:rmin+rest]
+			copy(b[rmin:], a[rmax:])
+			return b
+		}
+	}
+
+	return a[:rmin]
+}
+
+// search performs a binary search for UnixNano() v in a
+// and returns the position, i, where v would be inserted.
+// An additional check of a[i].UnixNano() == v is necessary
+// to determine if the value v exists.
+func (a {{.Name}}Values) search(v int64) int {
+	// Define: f(x) → a[x].UnixNano() < v
+	// Define: f(-1) == true, f(n) == false
+	// Invariant: f(lo-1) == true, f(hi) == false
+	lo := 0
+	hi := len(a)
+	for lo < hi {
+		mid := int(uint(lo+hi) >> 1)
+		if a[mid].UnixNano() < v {
+			lo = mid + 1 // preserves f(lo-1) == true
+		} else {
+			hi = mid // preserves f(hi) == false
+		}
+	}
+
+	// lo == hi
+	return lo
+}
+
+// FindRange returns the positions where min and max would be
+// inserted into the array. If a[0].UnixNano() > max or
+// a[len-1].UnixNano() < min then FindRange returns (-1, -1)
+// indicating the array is outside the [min, max].
+func (a {{.Name}}Values) FindRange(min, max int64) (int, int) {
+	minVal := a[0].UnixNano()
+	maxVal := a[len(a)-1].UnixNano()
+
+	if maxVal < min || minVal > max {
+		return -1, -1
+	}
+
+	return a.search(min), a.search(max)
 }
 
 // Include returns the subset values between min and max inclusive.